Update README with better instructions for installing from GitHub.
Some of Tarantula's dependencies live only on RubyForge, not on GitHub.
In order to have RubyGems automatically resolve and install all dependencies,
add GitHub as a RubyGems source, and then install Tarantula *without*
declaring a specific source for the installation.  Doing so allows RubyGems
to find relevance-tarantula on GitHub and find other dependencies that don't
live on GitHub (e.g., htmlentities) on RubyForge.